Added color feature for all components in the sense that: 1) The labels cannot entered anymore when the "tool" is selected. 2) The components that require a label will have a yellow background with blue text indicating the label requirement.
Prevented duplicated labeling in copy/paste actions, the behaviour is following:
a) Copy action => All labels are cleared on the clipboard and the paste action will return the selection with empty labels.
b) Cut action => All labels are copied to the clipboard, but, only the first paste action will return the selection with the "cutted" labels, a second paste will return the selection without labels.