Run vivado in batch mode
When running in tcl mode, Vivado will not close at the end of the
script. As the generated scripts have a 'exit' command at the end it not
a problem. But if a command in the tcl scipt failed, then vivado won't
exit and logisim will let the download progress bar open.

Fixed by replacing tcl mode with batch mode