Probe: Let the probe be in design as synthesisable, but do not synthesise it.
In this case, the probe should not be synthetised:
We could set the Probe as non-HDL element. But If we set the Probe
as non HDL element, logisim will not allow user to download the design.

In some case we need to use Logisim Simulation before running the design on the hardware.
During simulation, probes are very helpful to see signals values. And when simulation is ok,
the user does not want to delete all probes.
Thus, here we remove it form the netlist so it is transparent.